Part III: Cookies in pubs

+++ src/main/java/pl/pw/pamiw/biblio/controller/FileController.java
@@ -167,7 +167,20 @@ public class FileController {
 
     @RequestMapping(value = "/files/delete/{filename}", method = RequestMethod.GET)
     public String deleteFile(HttpServletRequest request, HttpServletResponse response, @PathVariable String filename) {
-        fileService.deleteFile(filename);
-        return "redirect:/files";
+        ResponseEntity responseEntity = checkCookies(request, response);
+        if (responseEntity.getStatusCode().equals(HttpStatus.OK)) {
+            Cookie[] cookies = request.getCookies();
+            Cookie user = Arrays.stream(cookies).filter(cookie -> cookie.getName().equals("user")).findAny().orElse(null);
+
+            if (jwtService.canIDelete(createToken(user.getValue(), "list"), user.getValue())) {
+                fileService.deleteFile(filename);
+                return "redirect:/files";
+            } else {
+                System.out.println("Nie udało się autoryzować JWT");
+                return "forbidden";
+            }
+        } else {
+            return "forbidden";
+        }
     }
 }
